+++ game/g_save.cpp
@@ -248,9 +248,9 @@ void WriteField1( FILE *f, field_t *field, byte *base ) {
 	case F_LSTRING:
 	case F_GSTRING:
 	{
-		size_t len;
+		int len;
 		if ( *( char ** ) p )
-			len = strlen( *( char ** ) p ) + 1;
+			len = ( int ) strlen( *( char ** ) p ) + 1;
 		else
 			len = 0;
 		*( int * ) p = len;
